# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/config": {
      "post": {
        "operationId": "ConfigController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateConfigDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "创建成功"
          },
          "400": {
            "description": "请求参数错误"
          },
          "401": {
            "description": "未授权"
          },
          "403": {
            "description": "权限不足"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "创建配置",
        "tags": [
          "系统配置"
        ]
      },
      "get": {
        "operationId": "ConfigController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "获取成功"
          },
          "401": {
            "description": "未授权"
          },
          "403": {
            "description": "权限不足"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "获取所有配置",
        "tags": [
          "系统配置"
        ]
      },
      "patch": {
        "operationId": "ConfigController_updateAll",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "更新成功"
          },
          "400": {
            "description": "请求参数错误"
          },
          "401": {
            "description": "未授权"
          },
          "403": {
            "description": "权限不足"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "更新所有配置",
        "tags": [
          "系统配置"
        ]
      }
    },
...
    "/upload/config": {
      "get": {
        "operationId": "UploadController_getUploadConfig",
        "parameters": [],
        "responses": {
          "200": {
            "description": "获取成功"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "获取上传配置",
        "tags": [
          "上传管理"
        ]
      }
    },
    "/upload/file": {
      "post": {
        "operationId": "UploadController_uploadFile",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "上传文件成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Upload"
                  }
                }
              }
            }
          },
          "400": {
            "description": "请求参数错误"
          },
          "403": {
            "description": "权限不足"
          },
          "500": {
            "description": "服务器错误"
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "bearer": []
          }
        ],
        "summary": "上传文件",
        "tags": [
          "上传管理"
        ]
      }
    },
    "/upload/info/{id}": {
      "get": {
        "operationId": "UploadController_getFileInfo",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "文件ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "获取文件信息成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Upload"
                }
              }
            }
          },
          "404": {
            "description": "文件不存在"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "获取文件信息",
        "tags": [
          "上传管理"
        ]
      }
    },
    "/upload": {
      "get": {
        "operationId": "UploadController_findAll",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "页码",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "每页数量",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 10,
              "type": "number"
            }
          },
          {
            "name": "sortBy",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sortOrder",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "获取所有上传文件成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Upload"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "bearer": []
          }
        ],
        "summary": "获取所有上传文件",
        "tags": [
          "上传管理"
        ]
      }
    },
    "/upload/migration/stats": {
      "get": {
        "operationId": "UploadController_getMigrationStats",
        "parameters": [],
        "responses": {
          "200": {
            "description": "获取成功"
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "bearer": []
          }
        ],
        "summary": "获取压缩处理统计（管理员）",
        "tags": [
          "上传管理"
        ]
      }
    },
    "/upload/migration/process": {
      "post": {
        "operationId": "UploadController_processUncompressedImages",
        "parameters": [
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "dryRun",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "limit": {
                    "type": "number",
                    "description": "处理数量限制",
                    "default": 100
                  },
                  "dryRun": {
                    "type": "boolean",
                    "description": "模拟运行（不实际处理）",
                    "default": false
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "处理完成"
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "bearer": []
          }
        ],
        "summary": "批量处理未压缩图片（管理员）",
        "tags": [
          "上传管理"
        ]
      }
    },
    "/upload/migration/metadata": {
      "post": {
        "operationId": "UploadController_fillMissingMetadata",
        "parameters": [
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "处理完成"
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "bearer": []
          }
        ],
        "summary": "补充图片元数据（管理员）",
        "tags": [
          "上传管理"
        ]
      }
    },
...
    "/report": {
      "post": {
        "operationId": "ReportController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateReportDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "创建成功"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "创建举报",
        "tags": [
          "举报管理"
        ]
      },
      "get": {
        "operationId": "ReportController_findAll",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "页码",
            "schema": {
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "每页数量",
            "schema": {
              "default": 10,
              "type": "number"
            }
          },
          {
            "name": "type",
            "required": false,
            "in": "query",
            "description": "举报类型",
            "schema": {
              "type": "string",
              "enum": [
                "USER",
                "ARTICLE",
                "COMMENT"
              ]
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "description": "处理状态",
            "schema": {
              "type": "string",
              "enum": [
                "PENDING",
                "PROCESSING",
                "RESOLVED",
                "REJECTED"
              ]
            }
          },
          {
            "name": "category",
            "required": false,
            "in": "query",
            "description": "举报分类",
            "schema": {
              "type": "string",
              "enum": [
                "SPAM",
                "ABUSE",
                "INAPPROPRIATE",
                "COPYRIGHT",
                "OTHER"
              ]
            }
          },
          {
            "name": "reporterId",
            "required": false,
            "in": "query",
            "description": "举报人ID",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "keyword",
            "required": false,
            "in": "query",
            "description": "关键词搜索（举报原因）",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sortBy",
            "required": false,
            "in": "query",
            "description": "排序字段",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sortOrder",
            "required": false,
            "in": "query",
            "description": "排序方向",
            "schema": {
              "type": "string",
              "enum": [
                "ASC",
                "DESC"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "获取成功"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "获取举报列表",
        "tags": [
          "举报管理"
        ]
      }
    },
    "/report/statistics": {
      "get": {
        "operationId": "ReportController_getStatistics",
        "parameters": [],
        "responses": {
          "200": {
            "description": "获取成功"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "获取举报统计",
        "tags": [
          "举报管理"
        ]
      }
    },
    "/report/{id}": {
      "get": {
        "operationId": "ReportController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "获取成功"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "获取举报详情",
        "tags": [
          "举报管理"
        ]
      },
      "patch": {
        "operationId": "ReportController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateReportDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "更新成功"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "更新举报状态",
        "tags": [
          "举报管理"
        ]
      },
      "delete": {
        "operationId": "ReportController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "删除成功"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "删除举报记录",
        "tags": [
          "举报管理"
        ]
      }
    },
...
  },
  "info": {
    "title": "PicArt API",
    "description": "PicArt 图片社区 API 文档",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateConfigDto": {
        "type": "object",
        "properties": {}
      },
...
      "Upload": {
        "type": "object",
        "properties": {}
      },
...
      "CreateReportDto": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "description": "举报类型",
            "enum": [
              "USER",
              "ARTICLE",
              "COMMENT"
            ]
          },
          "reason": {
            "type": "string",
            "description": "举报原因"
          },
          "category": {
            "type": "string",
            "description": "举报分类",
            "enum": [
              "SPAM",
              "ABUSE",
              "INAPPROPRIATE",
              "COPYRIGHT",
              "OTHER"
            ]
          },
          "description": {
            "type": "string",
            "description": "详细描述"
          },
          "reportedUserId": {
            "type": "number",
            "description": "被举报用户ID"
          },
          "reportedArticleId": {
            "type": "number",
            "description": "被举报文章ID"
          },
          "reportedCommentId": {
            "type": "number",
            "description": "被举报评论ID"
          }
        },
        "required": [
          "type",
          "reason",
          "category"
        ]
      },
      "UpdateReportDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "description": "处理状态",
            "enum": [
              "PENDING",
              "PROCESSING",
              "RESOLVED",
              "REJECTED"
            ]
          },
          "action": {
            "type": "string",
            "description": "处理动作",
            "enum": [
              "DELETE_CONTENT",
              "BAN_USER",
              "WARNING",
              "NONE"
            ],
            "example": "DELETE_CONTENT"
          },
          "result": {
            "type": "string",
            "description": "处理结果"
          }
        }
      },
...
    }
  }
}